Renamed folders in connector-thegraph
