+++ packages/connector-thegraph/src/org/parsers/repos.ts
@@ -1,4 +1,4 @@
-import { Repo as RepoDataGql } from "../graphql/types";
+import { Repo as RepoDataGql } from "../queries/types";
 import { RepoData } from "plumbery-core";
 
 export function parseRepo(
