+++ packages/connector-thegraph/src/org/queries/fragments.ts
