+++ packages/connector-thegraph/src/org/queries/index.ts
