+++ packages/connector-thegraph/src/org/queries/types.ts
