+++ packages/connector-thegraph/src/voting/index.ts
@@ -1,0 +1,11 @@
+import { GraphQLWrapper } from "..";
+
+export default class VotingConnectorTheGraph extends GraphQLWrapper {
+  // async votesForApp(): Vote[] {
+
+  // }
+
+  // async castsForVote(): Cast[] {
+
+  // }
+}
